requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
python-decouple>=3.8
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
django-silk>=5.1.0
djoser>=2.2.0
psycopg2-binary>=2.9.9
